% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_functions.R
\name{print.rankMANOVA}
\alias{print.rankMANOVA}
\title{Display rankMANOVA object}
\usage{
\method{print}{rankMANOVA}(x, ...)
}
\arguments{
\item{x}{A rankMANOVA object}

\item{...}{Additional parameters (currently not used)}
}
\description{
Returns a short summary of the results (test statistic with p-values)
}
