% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rankMANOVA-function.R
\name{rankMANOVA}
\alias{rankMANOVA}
\title{Rank-based Tests for Multivariate Data in Semi-Parametric Factorial Designs}
\usage{
rankMANOVA(formula, data, iter = 10000, alpha = 0.05, CPU, seed)
}
\arguments{
\item{formula}{A model \code{\link{formula}} object. The left hand side
contains the response variable and the right hand side contains the factor
variables of interest. An interaction term must be specified.}

\item{data}{A data.frame, list or environment containing the variables in
\code{formula}. Data must be in long format.}

\item{iter}{The number of iterations used for calculating the resampled
statistic. The default option is 10,000.}

\item{alpha}{A number specifying the significance level; the default is 0.05.}

\item{CPU}{The number of cores used for parallel computing. If omitted, cores are
detected via \code{\link[parallel]{detectCores}}.}

\item{seed}{A random seed for the resampling procedure. If omitted, no
reproducible seed is set.}

\item{subject}{The column name of the subjects in the data.}

\item{resampling}{The resampling method to be used, one of "paramBS"
(parametric bootstrap approach) and "WildBS" (wild bootstrap approach with
Rademacher weights). The Wild Bootstrap is calculated for all test statistics.}

\item{nested.levels.unique}{A logical specifying whether the levels of the nested factor(s)
are labeled uniquely or not. Default is FALSE, i.e., the levels of the nested
factor are the same for each level of the main factor.}
}
\value{
A \code{MANOVA} object containing the following components:
  \item{Descriptive}{Some descriptive statistics of the data for all factor
  level combinations. Displayed are the number of individuals per factor
  level combination and the vector of means (one column per dimension).}
  \item{Covariance}{The estimated covariance matrix.}
  \item{WTS}{The value of the WTS along with degrees of freedom of the
  central chi-square distribution and p-value.}
  \item{ATS}{The value of the
  ATS, degrees of freedom of the central F distribution and the corresponding
  p-value.}
  \item{MATS}{The value of the MATS.}
  \item{resampling}{p-values for the test statistic based on the
  chosen resampling approach.}
}
\description{
The rankMANOVA function calculates the ANOVA-type
statistic (ATS) with bootstrap p-values.
}
\details{

}
\section{}{

}
\examples{

}
\references{
Konietschke, F., Bathke, A. C., Harrar, S. W. and Pauly, M. (2015).
  Parametric and nonparametric bootstrap methods for general MANOVA. Journal
  of Multivariate Analysis, 140, 291-301.

  Friedrich, S., Brunner, E. and Pauly, M. (2017). Permuting longitudinal data
  in spite of the dependencies. Journal of Multivariate Analysis, 153, 255-265.

   Bathke, A., Friedrich, S., Konietschke, F., Pauly, M., Staffen, W., Strobl, N. and Hoeller, Y. (2016).
   Using EEG, SPECT, and Multivariate Resampling Methods
   to Differentiate Between Alzheimer's and other Cognitive Impairments. arXiv preprint arXiv:1606.09004.

  Friedrich, S., Konietschke, F., Pauly, M. (2016). GFD - An
  R-package for the Analysis of General Factorial Designs. Accepted for publication in
  Journal of Statistical Software.

  Friedrich, S., and Pauly, M. (2017). MATS: Inference for potentially singular and
  heteroscedastic MANOVA. arXiv preprint arXiv:1704.03731.
}
\seealso{
\code{\link{MANOVA.RM}}
}

