% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_functions.R
\name{summary.rankMANOVA}
\alias{summary.rankMANOVA}
\title{Summarizing a rankMANOVA object}
\usage{
\method{summary}{rankMANOVA}(object, ...)
}
\arguments{
\item{object}{A rankMANOVA object}

\item{...}{Additional parameters (currently not used)}
}
\description{
Returns a summary of the results including sample sizes and relative treatment effects for all groups as well
as the test statistic with resampling-based p-values
}
