% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/posthoc.R
\name{univariate}
\alias{univariate}
\title{Univariate post-hoc comparisons in nonparametric multivariate factorial designs}
\usage{
univariate(object, factor, data, ...)
}
\arguments{
\item{object}{A \code{rankMANOVA} object.}

\item{factor}{The factor for which univariate comparisons are desired. Must be one of the factors
used in the main analysis, of course.}

\item{data}{The data set to be used for the analysis. If none is specified, the data used for fitting
\code{object} is re-used.}

\item{...}{Not used yet.}
}
\value{
p-values for the univariate post-hoc comparisons of the chosen factor.
}
\description{
Univariate post-hoc comparisons in nonparametric multivariate factorial designs
}
\details{
The univariate function computes p-values for univariate comparisons.
Details on the tests can be found in Dobler, Friedrich and Pauly (2019).
Note that due to the formulation of our effect size vectors, these tests can be performed by applying
the closed testing principle, i.e., no alpha-corretion is needed.
}
\references{
Dobler, D., Friedrich, S., and Pauly, M. (2019).
             Nonparametric MANOVA in meaningful effects. Annals of the Institute of Statistical Mathematics.
}
