% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/posthoc.R
\name{pairwise}
\alias{pairwise}
\title{Pairwise post-hoc comparisons in nonparametric multivariate factorial designs}
\usage{
pairwise(object, type = NULL, base = 1, factor, uni = FALSE, ...)
}
\arguments{
\item{object}{A \code{rankMANOVA} object.}

\item{type}{The type of the pairwise comparison must be specified here.
Calculation is based on the contrMat function in package multcomp, see the corresponding help page
for details on the types of contrasts available.}

\item{base}{an interger specifying which group is considered the baseline group
for Dunnett contrasts, see \code{\link[multcomp]{contrMat}}.}

\item{factor}{The factor for which pairwise comparisons are desired. Must be one of the factors
used in the main analysis, of course.}

\item{uni}{Logical: should univariate comparisons also be computed? Default is FALSE.}

\item{...}{Not used yet.}
}
\value{
p-values for the multivariate and (if desired) univariate pairwise comparisons of the chosen factor.
}
\description{
Pairwise post-hoc comparisons in nonparametric multivariate factorial designs
}
\details{
The pairwise function computes p-values for pairwise comparisons as provided by
 \code{\link[multcomp]{contrMat}}. For the nonparametric comparisons,
 only Tukey's all-pairwise and Dunnett's many-to-one comparisons are used.
Note that due to the formulation of our effect size vectors, these tests can be performed by applying
the closed testing principle, i.e., no alpha-corretion is needed, see Dobler, Friedrich and Pauly (2019)
for details.
}
\references{
Dobler, D., Friedrich, S., and Pauly, M. (2019).
             Nonparametric MANOVA in meaningful effects. Annals of the Institute of Statistical Mathematics.
}
\seealso{
\code{\link[multcomp]{contrMat}}
}
