% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rankMANOVA-function.R
\name{rankMANOVA}
\alias{rankMANOVA}
\title{Rank-based Tests for Multivariate Data in Nonparametric Factorial Designs}
\usage{
rankMANOVA(
  formula,
  data,
  iter = 10000,
  alpha = 0.05,
  para = FALSE,
  CPU,
  dec = 3,
  seed,
  resampling = "bootstrap",
  nested.levels.unique = FALSE
)
}
\arguments{
\item{formula}{A model \code{\link{formula}} object. The left hand side
contains the response variables and the right hand side contains the factor
variables of interest. Data must be
provided in wide format.}

\item{data}{A data.frame containing the variables in
\code{formula}.}

\item{iter}{The number of iterations used for calculating the resampled
statistic. The default option is 10,000.}

\item{alpha}{A number specifying the significance level; the default is 0.05.}

\item{para}{Logical: should parallel computing be used? Default is FALSE.}

\item{CPU}{The number of cores used for parallel computing. If omitted, cores are
detected via \code{\link[parallel]{detectCores}}.}

\item{dec}{Number of decimals the results should be rounded to. Default is 3.}

\item{seed}{A random seed for the resampling procedure. If omitted, no
reproducible seed is set.}

\item{resampling}{The resampling method to be used, one of "bootstrap"
(sample-specific bootstrap approach) and "WildBS" (wild bootstrap approach with
Rademacher weights). The default is "WildBS".}

\item{nested.levels.unique}{A logical specifying whether the levels of the nested factor(s)
are labeled uniquely or not. Default is FALSE, i.e., the levels of the nested
factor are the same for each level of the main factor. For an example and more explanations
see the GFD package and the corresponding vignette.}
}
\value{
A \code{rankMANOVA} object containing the following components:
\item{Descriptive}{Some descriptive statistics of the data for all factor
  level combinations. Displayed are the number of individuals per factor
  level combination and the unweighted treatment effects for each dimension.}
  \item{Test}{The test statistic(s) and p-value(s) based on the
  chosen bootstrap approach.}
}
\description{
The rankMANOVA function calculates an ANOVA-type
statistic (ATS) with (wild) bootstrap p-values for nonparametric factorial designs with
multivariate data.
}
\details{
Implemented is an ANOVA-type test statistic for testing hypotheses formulated in Mann-Whitney-type
 effects in nonparametric factorial designs. Statistical inference is based on a wild or a sample-specific
 bootstrap approach. The unweighted treatment effects considered do not depend on sample sizes and allow for
 transitive ordering. The package thus provides an extension of the univariate \code{\link[rankFD]{rankFD}}
 package to multivariate data.
}
\section{NOTE}{
 The number of bootstrap iterations has been set to 100 in the examples due
to runtime restrictions on CRAN. Usually it is recommended to use at least 1000 iterations.
}

\examples{
 data("marketing")
 mymar <- marketing[, c("Sex", "Income", "Edu")]
 mymar2 <- na.omit(mymar)
 test <- rankMANOVA(cbind(Income, Edu) ~ Sex, data = mymar2, iter=100,
  resampling = "WildBS", CPU = 1)
 summary(test)


}
\references{
Dobler, D., Friedrich, S., and Pauly, M. (2017). Nonparametric MANOVA in Mann-Whitney effects.
}
\seealso{
\code{\link[rankFD]{rankFD}}
}
